+++ mods/BrewLAN_Units/BrewResearch/units/SAB1101/SAB1101_script.lua
@@ -1,0 +1,6 @@
+local TEnergyCreationUnit = import('/lua/defaultunits.lua').WindEnergyCreationUnit
+
+SAB1101 = Class(TEnergyCreationUnit) {
+}
+
+TypeClass = SAB1101