+++ mods/BrewLAN_Units/BrewResearch/units/SEB1101/SEB1101_script.lua
@@ -1,0 +1,6 @@
+local TEnergyCreationUnit = import('/lua/defaultunits.lua').WindEnergyCreationUnit
+
+SEB1101 = Class(TEnergyCreationUnit) {
+}
+
+TypeClass = SEB1101