+++ mods/BrewLAN/documentation/Changelog.md
@@ -829,6 +829,18 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   * R&D: Added notes for what each research item is built by.
   * R&D: Added icons.
 
+  ### 2018-03-31
+
+  * R&D: Added a UEF T1 wind turbine.
+
+  ### 2018-01-03
+
+  * R&D: Added random wind.
+  * R&D: Added an Aeon T1 wind turbine.
+  * R&D: Moved class definitions to defaultunits.lua
+  * R&D: Added icon for UEF wind turbine.
+  * R&D: Research locked normal UEF power generator.
+
 ## 2016-12-15|0.7.5.5
 
 * Release version.